feat: remove dice.yml check in pipeline
