+++ modules/pipeline/services/pipelinesvc/precheck.go
@@ -15,21 +15,11 @@
 package pipelinesvc
 
 import (
-	"strconv"
-	"strings"
-
-	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
-
 	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/modules/pipeline/commonutil/thirdparty/gittarutil"
 	"github.com/erda-project/erda/modules/pipeline/precheck"
-	"github.com/erda-project/erda/modules/pipeline/precheck/checkers/actionchecker/release"
 	"github.com/erda-project/erda/modules/pipeline/precheck/prechecktype"
 	"github.com/erda-project/erda/modules/pipeline/services/apierrors"
 	"github.com/erda-project/erda/modules/pipeline/spec"
-	"github.com/erda-project/erda/pkg/discover"
-	"github.com/erda-project/erda/pkg/parser/diceyml"
 	"github.com/erda-project/erda/pkg/parser/pipelineyml"
 )
 
@@ -48,21 +38,6 @@ func (s *PipelineSvc) PreCheck(pipelineYml *pipelineyml.PipelineYml, p *spec.Pip
 		Envs:                      p.Snapshot.Envs,
 		ClusterName:               p.ClusterName,
 		GlobalSnippetConfigLabels: p.Labels,
-	}
-
-	// files
-	if p.CommitDetail.RepoAbbr != "" {
-		diceymlByte, err := gittarutil.NewRepo(discover.Gittar(), p.CommitDetail.RepoAbbr).
-			FetchFile(p.GetCommitID(), "dice.yml", userID)
-		if err == nil {
-			itemsForCheck.Files["dice.yml"] = string(diceymlByte)
-		} else {
-			logrus.Error(err)
-		}
-	}
-	err = setItemForCheckRealDiceYml(p, &itemsForCheck, userID)
-	if err != nil {
-		return err
 	}
 
 	// 从 extension marketplace 获取 action
@@ -135,123 +110,6 @@ func (s *PipelineSvc) PreCheck(pipelineYml *pipelineyml.PipelineYml, p *spec.Pip
 		if err := s.dbClient.StoreAnalyzedCrossCluster(p.ID, analyzedCrossCluster); err != nil {
 			return apierrors.ErrPreCheckPipeline.InternalError(err)
 		}
-	}
-
-	return nil
-}
-
-// 用户可能在 release 中设置了 dice_development_yml,dice_test_yml,dice_staging_yml,dice_production_yml 等不同环境的 dice.yml, 但是对应的校验也要转化
-func setItemForCheckRealDiceYml(p *spec.Pipeline, itemForCheck *prechecktype.ItemsForCheck, userID string) error {
-	if p == nil {
-		return nil
-	}
-
-	y, err := pipelineyml.New([]byte(p.PipelineYml))
-	if err != nil {
-		return err
-	}
-	var worn error
-	// 遍历 pipeline yml 中的 action
-	y.Spec().LoopStagesActions(func(stage int, action *pipelineyml.Action) {
-		if action.Type != release.ActionType {
-			return
-		}
-		// 拿到 release 中指名的 yml
-		var realDiceYmlParam interface{}
-		workspace := p.Extra.DiceWorkspace
-		switch workspace {
-		case "DEV":
-			realDiceYmlParam = action.Params["dice_development_yml"]
-		case "PROD":
-			realDiceYmlParam = action.Params["dice_production_yml"]
-		case "TEST":
-			realDiceYmlParam = action.Params["dice_test_yml"]
-		case "STAGING":
-			realDiceYmlParam = action.Params["dice_staging_yml"]
-		default:
-			realDiceYmlParam = action.Params["dice_yml"]
-		}
-
-		if realDiceYmlParam == nil {
-			return
-		}
-
-		if p.CommitDetail.RepoAbbr == "" {
-			return
-		}
-
-		var realDiceYmlStr, ok = realDiceYmlParam.(string)
-		if !ok {
-			return
-		}
-
-		realDiceYmlSplit := strings.Split(realDiceYmlStr, "/")
-		var length = len(realDiceYmlSplit)
-		if length < 1 {
-			return
-		}
-
-		realDiceYmlName := realDiceYmlSplit[length-1]
-		diceYmlByte, err := gittarutil.NewRepo(discover.Gittar(), p.CommitDetail.RepoAbbr).
-			FetchFile(p.GetCommitID(), realDiceYmlName, userID)
-		if err != nil {
-			logrus.Errorf("get workspace %v dice_yml error %v", workspace, err)
-			return
-		}
-
-		var check = true
-		if action.Params["check_diceyml"] != nil {
-			check, err = strconv.ParseBool(action.Params["check_diceyml"].(string))
-			if err != nil {
-				check = true
-			}
-		}
-
-		// 将 dice.yml 和 其他环境的 yml 合并下
-		yml, err := composeEnvYml(itemForCheck.Files["dice.yml"], check, string(diceYmlByte), workspace.String())
-		if err != nil {
-			logrus.Errorf("composeEnvYml dice.yml error: %v", err)
-			worn = err
-			return
-		}
-		itemForCheck.Files["dice.yml"] = yml
-	})
-
-	return worn
-}
-
-func composeEnvYml(diceYaml string, check bool, otherYaml string, workspace string) (string, error) {
-	d, err := diceyml.New([]byte(diceYaml), check)
-	if err != nil {
-		return "", errors.Wrap(err, "new parser failed")
-	}
-
-	switch workspace {
-	case string(apistructs.DevWorkspace):
-		err = composeYaml(d, "development", otherYaml)
-	case string(apistructs.TestWorkspace):
-		err = composeYaml(d, "test", otherYaml)
-	case string(apistructs.StagingWorkspace):
-		err = composeYaml(d, "staging", otherYaml)
-	case string(apistructs.ProdWorkspace):
-		err = composeYaml(d, "production", otherYaml)
-	}
-	if err != nil {
-		return "", errors.Wrap(err, "failed to compose diceyml")
-	}
-
-	return d.YAML()
-}
-
-func composeYaml(targetYml *diceyml.DiceYaml, env, envYmlFile string) error {
-	envYml, err := diceyml.New([]byte(envYmlFile), false)
-	if err != nil {
-		return err
-	}
-
-	err = targetYml.Compose(env, envYml)
-	if err != nil {
-		return err
 	}
 
 	return nil
